editor paste event result return ClipboardEvent (#192732)
* editor paste event result return ClipboardEvent

* Remove commented out code

* Cleanup `applyDefaultPasteHandler`

---------
